api-rest/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tensorflow==2.15.0
numpy==1.26.3
Pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
